requirements.txt
transformers
datasets
sentence-transformers
faiss-cpu
gradio
numpy
pandas
